TST: test_hybridmotor.py reorganized.
