+++ tests/unit/test_hybridmotor.py
@@ -23,21 +23,6 @@ grains_center_of_mass_position = -0.1
 oxidizer_tank_position = 0.3
 
 
-@patch("matplotlib.pyplot.show")
-def test_hybrid_motor_info(mock_show, hybrid_motor):
-    """Tests the HybridMotor.all_info() method.
-
-    Parameters
-    ----------
-    mock_show : mock
-        Mock of the matplotlib.pyplot.show function.
-    hybrid_motor : rocketpy.HybridMotor
-        The HybridMotor object to be used in the tests.
-    """
-    assert hybrid_motor.info() == None
-    assert hybrid_motor.all_info() == None
-
-
 def test_hybrid_motor_basic_parameters(hybrid_motor):
     """Tests the HybridMotor class construction parameters.
